requirements-langgraph.txt
# Phase 6: Multi-Agent Foundation Dependencies
# LangChain/LangGraph for agent orchestration

# Core LangChain/LangGraph
langchain>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Ollama integration
langchain-ollama>=0.2.0
ollama>=0.3.0

# Vector store & memory
chromadb>=0.5.0
sentence-transformers>=3.0.0

# FastAPI for API endpoints
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0

# Additional utilities
numpy>=1.26.0
pandas>=2.2.0

# Testing
pytest-asyncio>=0.24.0
httpx>=0.27.0
